Replace all doubles with floats

+++ examples/src/runner/egg.h
@@ -50,7 +50,7 @@ public:
 
     virtual void update()
     {
-        const double sv = 400;
+        const float sv = 400;
         if (state == State::Grounded)
         {
             if (!touchingGround && transform->getPosition().y() > bottom)
@@ -59,7 +59,7 @@ public:
             }
             if (input->isDown(SDLK_UP))
             {
-                velocity = {velocity.x(), 525};
+                velocity = {velocity.x(), 525.0f};
                 state = State::Airborn;
                 airTime = 0.0;
             }
@@ -69,7 +69,7 @@ public:
             airTime += time->deltaTime();
             if (airTime > 0.3)
             {
-                velocity = {velocity.x(), 0.0};
+                velocity = {velocity.x(), 0.0f};
                 state = State::Falling;
             }
         }
@@ -81,11 +81,11 @@ public:
             {
                 state = State::Grounded;
                 velocity = {velocity.x(), 0.0};
-                transform->setPosition({pos.x(), (double)bottom});
+                transform->setPosition({pos.x(), (float)bottom});
             }
         }
 
-        double x = 0.0;
+        float x = 0.0f;
         if (input->isDown(SDLK_LEFT))
         {
             x -= sv;
@@ -101,7 +101,7 @@ public:
 
     std::shared_ptr<Mino::IInputSystem> input;
     int bottom;
-    double height;
+    float height;
     bool touchingGround = false;
 
 private:
@@ -109,8 +109,8 @@ private:
     std::shared_ptr<Mino::ITimeService> time;
     Mino::Transform::TransformRef transform;
 
-    Mino::Vector2<double> velocity = {0, 0};
-    const double gravity = 50.0;
+    Mino::Vector2<float> velocity = {0, 0};
+    const float gravity = 50.0;
     State state = State::Grounded;
-    double airTime = 1000.0;
+    float airTime = 1000.0;
 };
