+++ src/common/bounding_box/bounding_box.h
@@ -8,11 +8,11 @@ namespace Mino
 class BoundingBox
 {
 public:
-    BoundingBox(Vector2<double> center, double halfDimension)
+    BoundingBox(Vector2<float> center, float halfDimension)
         : center(center), halfWidth(halfDimension), halfHeight(halfDimension)
     {
     }
-    BoundingBox(Vector2<double> center, double width, double height)
+    BoundingBox(Vector2<float> center, float width, float height)
         : center(center), halfWidth(width * 0.5), halfHeight(height * 0.5)
     {
     }
@@ -23,21 +23,21 @@ public:
     BoundingBox& operator=(BoundingBox const&) = default;
     BoundingBox& operator=(BoundingBox&&) = default;
 
-    bool containsPoint(Vector2<double> const& point) const;
+    bool containsPoint(Vector2<float> const& point) const;
     bool intersects(BoundingBox const& other) const;
 
-    Vector2<double> getCenter() const { return center; }
-    double getWidth() { return halfWidth * 2; }
-    double getHeight() { return halfHeight * 2; }
+    Vector2<float> getCenter() const { return center; }
+    float getWidth() { return halfWidth * 2; }
+    float getHeight() { return halfHeight * 2; }
 
-    void setCenter(Vector2<double> const& value) { center = value; }
-    void setWidth(double value) { halfWidth = value * 0.5; }
-    void setHeight(double value) { halfHeight = value * 0.5; }
+    void setCenter(Vector2<float> const& value) { center = value; }
+    void setWidth(float value) { halfWidth = value * 0.5; }
+    void setHeight(float value) { halfHeight = value * 0.5; }
 
 private:
-    Vector2<double> center;
-    double halfWidth;
-    double halfHeight;
+    Vector2<float> center;
+    float halfWidth;
+    float halfHeight;
 };
 
 } // namespace Mino
