+++ src/component/sprite_render/sprite_render.cpp
@@ -18,7 +18,7 @@ void SpriteRenderComponent::update()
 
 void SpriteRenderComponent::render(Transform::TransformRef camera)
 {
-    auto offset = camera ? camera->getPosition() : Vector2<double>{0, 0};
+    auto offset = camera ? camera->getPosition() : Vector2<float>{0, 0};
     auto position = transform->getPosition() - offset;
     auto x = static_cast<int>(position.x());
     auto y = static_cast<int>(position.y());
