+++ src/engine/scene/scene.h
@@ -29,7 +29,7 @@ public:
     virtual void start() {}
 
     template <typename... TComponents>
-    std::shared_ptr<GameObject> createGameObject(Vector2<double> position = {0, 0});
+    std::shared_ptr<GameObject> createGameObject(Vector2<float> position = {0, 0});
     void destroyGameObject(std::shared_ptr<GameObject>);
 
     IEngineCore* getEngineCore() { return engine; }
@@ -65,7 +65,7 @@ template <typename T, typename... Ts> void Scene::addComponentHelper(GameObject&
 }
 
 template <typename... TComponents>
-std::shared_ptr<GameObject> Scene::createGameObject(Vector2<double> position)
+std::shared_ptr<GameObject> Scene::createGameObject(Vector2<float> position)
 {
     auto go = createEmptyGameObject();
     auto tr = go->getTransform();
